[Refactor] primary, secondary 버튼을 CustomTextButton 팩토리 생성자로 통합 #6
